secondtask/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
backoff==2.2.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Django==5.1.4
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-API==3.1
hiredis==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
python-stdnum==1.20
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
Werkzeug==3.1.3
